Refactored packages and deleted unnecessary
