+++ app/src/main/java/com/therxmv/telegramthemer/data/source/SharedPrefsDataSource.kt
@@ -1,13 +1,17 @@
 package com.therxmv.telegramthemer.data.source
 
 import com.tencent.mmkv.MMKV
-import com.therxmv.telegramthemer.ui.editor.data.ThemeState
-import com.therxmv.telegramthemer.utils.SHARED_PREFS_THEME_KEY
+import com.therxmv.telegramthemer.domain.model.ThemeState
+import com.therxmv.telegramthemer.domain.source.SharedPrefsSource
 import javax.inject.Inject
 
 class SharedPrefsDataSource @Inject constructor(
     private val mmkv: MMKV,
 ) : SharedPrefsSource {
+
+    companion object {
+        private const val SHARED_PREFS_THEME_KEY = "ThemeStateKey"
+    }
 
     override fun saveThemeState(themeState: ThemeState) {
         mmkv.encode(SHARED_PREFS_THEME_KEY, themeState)
