+++ app/src/main/java/com/therxmv/telegramthemer/data/theme/utils/TintsExtensions.kt
@@ -1,4 +1,4 @@
-package com.therxmv.telegramthemer.ui.editor.data.utils
+package com.therxmv.telegramthemer.data.theme.utils
 
 /**
  * Generates list of color tints from 0 to 10 luminosity. 0 is black and 10 is white
