+++ app/src/main/java/com/therxmv/telegramthemer/domain/adapter/PreviewColorsAdapter.kt
@@ -1,0 +1,8 @@
+package com.therxmv.telegramthemer.domain.adapter
+
+import com.therxmv.preview.PreviewColorsModel
+import com.therxmv.telegramthemer.domain.model.ThemeState
+
+interface PreviewColorsAdapter {
+    fun getThemePreviewColors(themeState: ThemeState): PreviewColorsModel
+}
