+++ app/src/main/java/com/therxmv/telegramthemer/domain/usecase/GetAtthemeFileUseCase.kt
@@ -1,0 +1,14 @@
+package com.therxmv.telegramthemer.domain.usecase
+
+import com.therxmv.telegramthemer.domain.adapter.ThemeFileAdapter
+import com.therxmv.telegramthemer.domain.model.ThemeState
+import java.io.File
+import javax.inject.Inject
+
+class GetAtthemeFileUseCase @Inject constructor(
+    private val themeFileAdapter: ThemeFileAdapter,
+) {
+
+    operator fun invoke(themeState: ThemeState): File =
+        themeFileAdapter.createThemeFile(themeState)
+}
