+++ app/src/main/java/com/therxmv/telegramthemer/domain/usecase/GetPreviewColorsModelUseCase.kt
@@ -1,0 +1,14 @@
+package com.therxmv.telegramthemer.domain.usecase
+
+import com.therxmv.preview.PreviewColorsModel
+import com.therxmv.telegramthemer.domain.adapter.PreviewColorsAdapter
+import com.therxmv.telegramthemer.domain.model.ThemeState
+import javax.inject.Inject
+
+class GetPreviewColorsModelUseCase @Inject constructor(
+    private val previewColorsAdapter: PreviewColorsAdapter,
+) {
+
+    operator fun invoke(themeState: ThemeState): PreviewColorsModel =
+        previewColorsAdapter.getThemePreviewColors(themeState)
+}
