+++ app/src/main/java/com/therxmv/telegramthemer/ui/editor/ThemeStateListener.kt
@@ -1,6 +1,6 @@
 package com.therxmv.telegramthemer.ui.editor
 
-import com.therxmv.telegramthemer.ui.editor.data.ThemeState
+import com.therxmv.telegramthemer.domain.model.ThemeState
 
 interface ThemeStateListener {
     fun onStateChange(themeState: ThemeState)
