--- app/src/main/java/com/therxmv/telegramthemer/ui/editor/data/preview/PreviewColorsAdapter.kt
@@ -1,8 +1,0 @@
-package com.therxmv.telegramthemer.ui.editor.data.preview
-
-import com.therxmv.preview.PreviewColorsModel
-import com.therxmv.telegramthemer.ui.editor.data.ThemeState
-
-interface PreviewColorsAdapter {
-    fun getDefaultThemeColors(themeState: ThemeState): PreviewColorsModel
-}
