+++ app/src/main/java/com/therxmv/telegramthemer/ui/editor/options/MoreOptionsSubscriber.kt
@@ -1,6 +1,6 @@
 package com.therxmv.telegramthemer.ui.editor.options
 
-import com.therxmv.telegramthemer.ui.editor.data.ThemeState
+import com.therxmv.telegramthemer.domain.model.ThemeState
 
 interface MoreOptionsSubscriber {
     fun onPropertyChange(themeState: ThemeState)
