--- app/src/main/java/com/therxmv/telegramthemer/utils/Constants.kt
@@ -1,43 +1,0 @@
-package com.therxmv.telegramthemer.utils
-
-const val DEFAULT_COLOR = "299fe9"
-const val SHARED_PREFS_THEME_KEY = "ThemeStateKey"
-const val DEFAULT_HEX_COLOR = "#299FE9"
-
-object ThemeFileNames {
-    const val DEFAULT_DAY_THEME = "TheDay.attheme"
-    const val DEFAULT_DAY_TEMPLATE = "theday_template.attheme"
-
-    const val DEFAULT_NIGHT_THEME = "TheNight.attheme"
-    const val DEFAULT_AMOLED_THEME = "TheAmoled.attheme"
-    const val DEFAULT_NIGHT_TEMPLATE = "theday_dark_template.attheme"
-
-    const val SOZA_DAY_THEME = "Soza Day.attheme"
-    const val SOZA_DAY_TEMPLATE = "thesoza_template.attheme"
-
-    const val SOZA_NIGHT_THEME = "Soza Night.attheme"
-    const val SOZA_AMOLED_THEME = "Soza Amoled.attheme"
-    const val SOZA_NIGHT_TEMPLATE = "thesoza_dark_template.attheme"
-}
-
-object DefaultTheme {
-    const val DEFAULT_FACTOR_200 = 0.8
-    const val DEFAULT_FACTOR_300 = 0.2
-    const val DEFAULT_FACTOR_700 = 0.2
-    const val DEFAULT_FACTOR_800 = 0.4
-}
-
-object SozaTheme {
-    const val SOZA_FACTOR_200 = 0.8
-    const val SOZA_FACTOR_300 = 0.6
-    const val SOZA_FACTOR_700 = 0.3
-    const val SOZA_FACTOR_800 = 0.7
-}
-
-object ThemeAccent {
-    const val ACCENT_200 = "ac_200"
-    const val ACCENT_300 = "ac_300"
-    const val ACCENT_500 = "ac_500"
-    const val ACCENT_700 = "ac_700"
-    const val ACCENT_800 = "ac_800"
-}
