--- app/src/main/res/drawable/ic_twotone_preview_24.xml
@@ -1,15 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:tint="#000000"
-    android:viewportWidth="24"
-    android:viewportHeight="24">
-    <path
-        android:fillAlpha="0.3"
-        android:fillColor="@android:color/white"
-        android:pathData="M12,6.5c-3.79,0 -7.17,2.13 -8.82,5.5 1.65,3.37 5.02,5.5 8.82,5.5s7.17,-2.13 8.82,-5.5C19.17,8.63 15.79,6.5 12,6.5zM12,16.5c-2.48,0 -4.5,-2.02 -4.5,-4.5S9.52,7.5 12,7.5s4.5,2.02 4.5,4.5 -2.02,4.5 -4.5,4.5z"
-        android:strokeAlpha="0.3" />
-    <path
-        android:fillColor="@android:color/white"
-        android:pathData="M12,4.5C7,4.5 2.73,7.61 1,12c1.73,4.39 6,7.5 11,7.5s9.27,-3.11 11,-7.5c-1.73,-4.39 -6,-7.5 -11,-7.5zM12,17.5c-3.79,0 -7.17,-2.13 -8.82,-5.5C4.83,8.63 8.21,6.5 12,6.5s7.17,2.13 8.82,5.5c-1.65,3.37 -5.03,5.5 -8.82,5.5zM12,7.5c-2.48,0 -4.5,2.02 -4.5,4.5s2.02,4.5 4.5,4.5 4.5,-2.02 4.5,-4.5 -2.02,-4.5 -4.5,-4.5zM12,14.5c-1.38,0 -2.5,-1.12 -2.5,-2.5s1.12,-2.5 2.5,-2.5 2.5,1.12 2.5,2.5 -1.12,2.5 -2.5,2.5z" />
-</vector>
