--- app/src/main/res/layout/color_picker_layout.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<top.defaults.colorpicker.ColorPickerView
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    app:enableAlpha="false"
-    app:enableBrightness="true"/>
