--- app/src/main/res/layout/settings.xml
@@ -1,83 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:gravity="center"
-    android:layout_height="wrap_content"
-    android:layout_width="match_parent"
-    android:animateLayoutChanges="true"
-    android:orientation="vertical">
-
-    <com.google.android.material.textview.MaterialTextView
-        android:layout_height="wrap_content"
-        android:layout_width="match_parent"
-        android:text="@string/sett_title"
-        android:textAppearance="@style/TextAppearance.Material3.TitleLarge"
-        android:textStyle="bold" />
-
-    <com.google.android.material.textfield.TextInputLayout
-        style="@style/Widget.Material3.TextInputLayout.OutlinedBox.ExposedDropdownMenu"
-        android:id="@+id/styleDropdown"
-        android:layout_height="wrap_content"
-        android:layout_marginVertical="10dp"
-        android:layout_width="match_parent"
-        app:boxBackgroundColor="?attr/colorSurfaceVariant"
-        app:boxStrokeWidth="0dp"
-        app:boxStrokeWidthFocused="0dp">
-
-        <com.google.android.material.textfield.MaterialAutoCompleteTextView
-            android:id="@+id/styleDropdownItems"
-            android:inputType="none"
-            android:layout_height="match_parent"
-            android:layout_width="match_parent"
-            android:text="@string/sett_dropdown" />
-
-    </com.google.android.material.textfield.TextInputLayout>
-
-    <com.google.android.material.checkbox.MaterialCheckBox
-        android:id="@+id/cbDarkTheme"
-        android:layoutDirection="rtl"
-        android:layout_height="wrap_content"
-        android:layout_width="match_parent"
-        android:text="@string/sett_dark_theme"
-        android:textAppearance="@style/TextAppearance.Material3.BodyLarge" />
-
-    <com.google.android.material.checkbox.MaterialCheckBox
-        android:enabled="false"
-        android:id="@+id/cbAmoledTheme"
-        android:layoutDirection="rtl"
-        android:layout_height="wrap_content"
-        android:layout_width="match_parent"
-        android:text="@string/sett_amoled_theme"
-        android:textAppearance="@style/TextAppearance.Material3.BodyLarge" />
-
-    <com.google.android.material.checkbox.MaterialCheckBox
-        android:id="@+id/cbMonet"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layoutDirection="rtl"
-        android:text="@string/sett_monet"
-        android:textAppearance="@style/TextAppearance.Material3.BodyLarge"
-        android:visibility="gone"
-        tools:visibility="visible" />
-
-    <com.google.android.material.checkbox.MaterialCheckBox
-        android:id="@+id/cbMonetBg"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layoutDirection="rtl"
-        android:text="@string/sett_monet_bg"
-        android:textAppearance="@style/TextAppearance.Material3.BodyLarge"
-        android:visibility="gone"
-        tools:visibility="visible" />
-
-    <com.google.android.material.checkbox.MaterialCheckBox
-        android:id="@+id/cbGradient"
-        android:layoutDirection="rtl"
-        android:layout_height="wrap_content"
-        android:layout_width="match_parent"
-        android:text="@string/sett_gradient"
-        android:textAppearance="@style/TextAppearance.Material3.BodyLarge" />
-
-</LinearLayout>
