--- app/src/main/res/menu/menu_about.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <item
-        android:id="@+id/action_about"
-        android:icon="@drawable/ic_twotone_info_24"
-        android:title="About"
-        app:showAsAction="always"/>
-
-</menu>
