+++ app/src/main/res/values-v31/colors.xml
@@ -6,8 +6,6 @@
     <color name="theme_accent1_200">@android:color/system_accent1_200</color>
     <color name="theme_accent1_500">@android:color/system_accent1_500</color>
     <color name="theme_neutral1_50">@android:color/system_neutral1_50</color>
-    <color name="theme_neutral1_100">@android:color/system_neutral1_100</color>
     <color name="theme_neutral1_400">@android:color/system_neutral1_400</color>
-    <color name="theme_neutral1_800">@android:color/system_neutral1_800</color>
     <color name="theme_neutral1_900">@android:color/system_neutral1_900</color>
 </resources>
